Fix issue #1
Move validation improperly entered castling short circuit when the king is in check and any piece attempts to move from the king starting point to the endpoint of a castling move. Issue resolved by first testing whether the king is still in the starting position as a proxy for whether a castling move is being attempted.