Issue #13 allow short algebraic notation moves.
Add a San class to parse and store the internals of a short algebraic move.
Add san_to_long method to Game to convert the san move to the long version.
Add apply_san_move method to Game to apply the move

Replace xrange with range so that tests run in python3

Add find_all_pieces method to Board similar to find_piece, but returns a list of all indexes instead of just the first one.

Add tests to cover all the new code.